refactor: 모달창 열렸을 때 스크롤 제어 및 적용 버튼 수정
